feat(smartrouter): add --bench-after, the endpoint disable threshold
FAILOVER-TASKS section 6's `bench-after` row. Completes section 2: piece B taught
the counter to see node errors, and this makes the threshold settable.

MaxConsecutiveConnectionAttempts becomes a var, defaulting to DefaultBenchAfter
(50) and set once at startup before any relay is served. Its old comment is split:
the "raised from 5 to 50" history belongs to the default, not to the knob.

The name now means what section 2 says. Before piece B this flag would have
under-delivered on it — "N failed requests in a row, whatever the failures are"
while the counter was blind to a node error inside an HTTP 200, which is the
commonest shape of a dying node. That is why C landed last rather than first.

50 is inherited, not derived. It was chosen to protect a single-endpoint
deployment against transport blips, before the counter could see node errors at
all; the same number now covers a strictly larger set of failures. Kept as the
default so nothing moves on upgrade, and flagged in the comment as the thing to
revisit on production evidence.

Bound to viper, not only to the flag set. Without the binding
viper.GetUint64 reads 0 for every operator who configures it in config.yml — and
0 is exactly the value SetBenchAfter rejects, so the setting would silently do
nothing while appearing to work. This is the same class of defect flagged in
review on #262's resolveSelectionWeights, caught here before it shipped.

Zero falls back to the default with a warning rather than aborting the router,
matching MinSelectionChance and the optimizer weights. Zero would disable an
endpoint on its first failed request, which on a single-endpoint deployment
reproduces the "No pairings" symptom the default exists to avoid.

uint64 rather than int, matching Endpoint.ConnectionRefusals and
probeReenableTrialBudget — the values it is actually compared against. Untyped
constants adapted silently; a var cannot, so the type had to be chosen rather
than inherited. Call sites and test loop counters follow.

Tests: validation table including the zero fallback and the legal 1; that the
threshold really drives the disable rather than parsing and changing nothing;
that the budget stays CONSECUTIVE (twenty failures broken up by successes never
disable); that the probe trial budget tracks the threshold instead of a hardcoded
47; and that a threshold below the trial budget floors at zero rather than
wrapping — a guard that is easy to lose the moment a constant becomes a variable
an operator can set to 1.

Verification: go build, go vet, gofmt clean; full go test ./... green across
three consecutive runs. One earlier run reported a failure that did not reproduce
in those three and left no failing package in its output; this package has a
documented rare flake under full-suite load (TestConnectorGrpc, noted during the
MAG-2599 work), so it is recorded here rather than claimed clean.

+++ protocol/lavasession/direct_rpc_session_selection_test.go
@@ -229,7 +229,7 @@ func TestFetchEndpointConnection_DirectRPC_BackoffAndSelfHeal(t *testing.T) {
 
 	// (1) + (2a): one short of the threshold the endpoint is still enabled and is
 	// offered for relay despite the unreachable upstream.
-	for i := 0; i < MaxConsecutiveConnectionAttempts-1; i++ {
+	for i := uint64(0); i < MaxConsecutiveConnectionAttempts-1; i++ {
 		endpoint.MarkUnhealthy()
 	}
 	require.True(t, endpoint.Enabled, "endpoint must stay enabled below the consecutive-failure threshold")
